fix: validate all requet body and params using zod

+++ src/controllers/subject/deleteById.ts
@@ -3,10 +3,16 @@ import { Request, Response } from "express";
 import deleteById from "../../lib/subjects/deleteById";
 import badRequest from "@utils/badRequest";
 import notFound from "@utils/notFound";
+import { idParamSchema } from "@schemas/zod-schema";
 const deleteByIdController = async (req: Request, res: Response) => {
     try {
         const { id } = req.params;
-        //TODO: validate request body using Zod
+        // validate request body using Zod
+        const parsedParams = idParamSchema.safeParse(req.params);
+        if (!parsedParams.success) {
+            return badRequest(res, parsedParams.error.issues[0].message);
+        }
+
         if (!id) return badRequest(res, "Id is required");
 
         const subject = await deleteById(id);
